fix: DomainArticleImageを分離
